// src/enum.ts
export enum EColorClothes {
  BLACK = '/color/black.jpg',
  WHITE = '/color/white.jpg',
  RED = '/color/red.jpg',
}

export enum EImageBlackClothes {
  FRONT = '/images/black/front.jpg',
  BACK = '/images/black/back.jpg',
  LEFT = '/images/black/left.jpg',
  RIGHT = '/images/black/right.jpg',
  NECK = '/images/black/neck.jpg',
}

export enum EImageWhiteClothes {
  FRONT = '/images/white/front.jpg',
  BACK = '/images/white/back.jpg',
  LEFT = '/images/white/left.jpg',
  RIGHT = '/images/white/right.jpg',
  NECK = '/images/white/neck.jpg',
}

export enum EImageRedClothes {
  FRONT = '/images/red/front.jpg',
  BACK = '/images/red/back.jpg',
  LEFT = '/images/red/left.jpg',
  RIGHT = '/images/red/right.jpg',
  NECK = '/images/red/neck.jpg',
}

export enum EIdObjectFabric {
  VALID_AREA_FRONT_BACK = 'valid_area_front_back',
  GUIDE_BORDER_FRONT_BACK = 'guide_border_front_back',
  TEXT = 'text',
  IMAGE = 'image',
}